+++ api/api_helpers.py
@@ -1,13 +1,14 @@
+import faulthandler
+faulthandler.enable()  # will catch segfaults and write to STDERR
+
+from functools import cache
+from html import escape as html_escape
+import re
+import math
 import typing
 import ipaddress
 import json
 import uuid
-import faulthandler
-from functools import cache
-from html import escape as html_escape
-import re
-
-faulthandler.enable()  # will catch segfaults and write to STDERR
 
 from starlette.background import BackgroundTask
 from fastapi.responses import ORJSONResponse
@@ -816,3 +817,64 @@ def get_connecting_ip(request):
         return connecting_ip.split(",")[0]
 
     return request.client.host
+
+
+def replace_nan_with_zero(obj):
+    if isinstance(obj, dict):
+        for k, v in obj.items():
+            if isinstance(v, (dict, list)):
+                replace_nan_with_zero(v)
+            elif isinstance(v, float) and math.isnan(v):
+                obj[k] = 0
+    elif isinstance(obj, list):
+        for i, item in enumerate(obj):
+            if isinstance(item, (dict, list)):
+                replace_nan_with_zero(item)
+            elif isinstance(item, float) and math.isnan(item):
+                obj[i] = 0
+    return obj
+
+# Refactor have this in the Pydantic model?
+# https://github.com/green-coding-solutions/green-metrics-tool/issues/907
+def validate_hog_measurement_data(data):
+    required_top_level_fields = [
+        'coalitions', 'all_tasks', 'elapsed_ns', 'processor', 'thermal_pressure'
+    ]
+    for field in required_top_level_fields:
+        if field not in data:
+            raise ValueError(f"Missing required field: {field}")
+
+    # Validate 'coalitions' structure
+    if not isinstance(data['coalitions'], list):
+        raise ValueError("Expected 'coalitions' to be a list")
+
+    for coalition in data['coalitions']:
+        required_coalition_fields = [
+            'name', 'tasks', 'energy_impact_per_s', 'cputime_ms_per_s',
+            'diskio_bytesread', 'diskio_byteswritten', 'intr_wakeups', 'idle_wakeups'
+        ]
+        for field in required_coalition_fields:
+            if field not in coalition:
+                raise ValueError(f"Missing required coalition field: {field}")
+            if field == 'tasks' and not isinstance(coalition['tasks'], list):
+                raise ValueError(f"Expected 'tasks' to be a list in coalition: {coalition['name']}")
+
+    # Validate 'all_tasks' structure
+    if 'energy_impact_per_s' not in data['all_tasks']:
+        raise ValueError("Missing 'energy_impact_per_s' in 'all_tasks'")
+
+    # Validate 'processor' structure based on the processor type
+    processor_fields = data['processor'].keys()
+    if 'ane_energy' in processor_fields:
+        required_processor_fields = ['combined_power', 'cpu_energy', 'gpu_energy', 'ane_energy']
+    elif 'package_joules' in processor_fields:
+        required_processor_fields = ['package_joules', 'cpu_joules', 'igpu_watts']
+    else:
+        raise ValueError("Unknown processor type")
+
+    for field in required_processor_fields:
+        if field not in processor_fields:
+            raise ValueError(f"Missing required processor field: {field}")
+
+    # All checks passed
+    return True
